refactor: decompose InteractiveTutorial component into hooks and subcomponents
